Fix windows ci builds (#108)
Windows builds always used to work fine for me locally, because my
.gitconfig has `autocrlf = input`. But it seems as though there are some
line-endings in our fixtures that may not always be `lf` like they are
when we generate snapshots, presumably on Linux/Mac.

So this PR simply tells CI, hey autocrlf first. And yay, things pass!! I
also added a .gitattributes to force `lf` line-endings on all source
files.

Here is a sample of CI passing on Windows:
https://github.com/maraisr/hdx/actions/runs/15894577728/job/44823352507

closes: https://github.com/csskit/csskit/issues/66