Implements ordered combinators and uncomments some align properties (#113)
Implements Ordered Combinators def values and uncomments some align
properties as it can now be parsed with this addition.

It currently only supports optional types, ie `<color>?` and not
optional idents, like `auto?`. There are a handful of properties that
make use of optional idents in combinators, but want to handle that
separately.

There are probably more properties we can uncomment with this change,
but feel that can possibly be handled in a followup.