Refactor def.rs for better codegen (#142)
This PR attempts to clean up def.rs a little, with a few key techniques
that result in tidier codegen on the other side:

# Optimization 1

`GenerateParseImpl::parse_steps` used to take a `capture:
Option<TokenStream>` which could be used as a binding to store the
parsed values. Instead we now _return_ a tuple of 2 `TokenStream`s - the
first is the parse "steps" and the second is the parse "result"; an
expression that can be captured as a binding or enum variant.

The capture binding was overly restrictive on what parsing could do
because it could only store a single binding, meaning some values ended
up being stored as units, for example `<color>{1,2} | auto` would result
in `enum Foo { Color((Color, Option<Color>)), Auto(T![Ident]) }`. With
this change the same syntax can now produce `enum Foo { Color(Color,
Option<Color>), Auto(T![Ident]) }` - the difference being that the
`Color` enum variant is no longer a single tuple of 2 Colors, but
instead 2 values! This is important because it enables simpler struct
types.

The other benefit to this is that parse steps which are a single
expression can be packed into the "result" `TokenStream` (the second
entry in the tuple). This means single expression tokens no longer
creating a binding, only to return use that binding in the next line, so
for e.g. a step like:

```rs
let val = p.parse::<crate::CalcSize>()?;
Ok(Self::CalcSizeFunction(val))
```

Can be simplified to:

```rs
Ok(Self::CalcSizeFunction(p.parse::<crate::CalcSize>()?))
```

# Optimization 2

We used to check for `Def::Multiplier` and try to optimize away fixed or
bounded ranges. So for example `<bar>{1,2}` would optimize away to an
ordered combinator - so instead of `struct Foo<'a>(Vec<'a, Bar>)` you'd
end up with `struct Foo(Bar, Option<Bar>)`.

This is a very useful optimisation, not just for the cleanliness of
generated output but it also potentially avoids allocations.

The issue with the way we handled this optimisation is that it was done
on the code generation phase, meaning we had lots of repetative branches
to deal with it, and not all were covered - for example `<bar>{1,2} |
auto` would break the optimisation (because the code inside of
Combinator didn't include this optimisation case) and so while
`<bar>{1,2}` would optimise to `struct Foo<'a>(Vec<'a, Bar>)`,
`<bar>{1,2} | auto` would still be `enum Foo<'a> { Bar(Vec<'a, Bar>),
Auto(T![Ident]) }`. No good.

So, rather than putting this optimisation into the _code generation_
phase, it has been lifted to the _def parsing_ phase. When the parser
encounters a fixed or bounded range instead of outputting a
`Def::Multiplier`, it'll output a `Def::Combinator(_, Ordered)`. This
means that the code generation can forget about branching for these
cases and we have the added benefit of handling this at any depth level.

I think the snapshot tests speak to the improvements here.