enable cue-before, cue-after (#144)
On the hunt for more low hanging fruit I found `cue-after` and
`cue-before` which needed a small bit of work to be enabled:

- Both use `<decibel>` which is specified as the `db` dimension. Easy
enough, we just need to add the `db` dimension to the lexer (which
encodes all knwon units in an enum for efficiency), then as a token
macro (so it can be used), then in def.rs (so `<decibel>` is routed to
`T![Dimension:Db]`. Et voila.
- Both use `<uri>` which seems to be an archaic css2 alias of `<url>`.
Easy enough, we can simply add `Url` as a type to `DefType` and alias it
as `uri`.

With these two simple changes, we now have these two rules enabled!