enable text-emphasis-skip (#145)
`text-emphasis-skip` is an interesting value:

```
spaces || punctuation || symbols || narrow
```

This effectively says "it can be any of these keywords, in any order.
This is quite simple to define as a data-type:

```rs
struct TextEmphasisSkipStyleValue {
  spaces: Option<T![Ident]>,
  punctuation: Option<T![Ident]>,
  symbols: Option<T![Ident]>,
  narrow: Option<T![Ident]>,
}
```

The parsing rules should be farily evident from this struct. 

Sadly the soluton is a little unweildy (at least as I've naively done
it). We check especially for a Combinator of Optionals where all the
options are Idents, and branch if they are - generating the defintion
and code for this exactly.

Hopefully some more style values will leverage this technique so we
don't have all these code branches for one style value.