Prefer From<T> for Cursor (over <&T>), From<&T> for Span (over <T>) (#166)
This tries to clean up the `From<T> for Span` and `From<T> for Cursor`
types, to reduce the amount of boilerplate. As most things that can
`From<T> for Cursor` are likey to be stack allocated, then `From<T>` has
been chosen over `From<&T>` - and all `From<&T>` implementations have
been dropped, and their callsites cleaned up to deref. Conversely,
`From<&T> for Span` has been chosen because the aim is to let every AST
be `Into<Span>` and some AST nodes have to allocate on the heap, so it
would be bad to `.clone()` these just to get the span.