implement IntoCursor derive
This derive allows us to implement `From<T> for Cursor`, and also get `From<T> for Token` and `From<T> for Span` implementations for close to free. All that's required is a `derive(IntoCursor)` on a struct or enum.
This removes a fair amount of code but more importantly consistently implements the three pairs across any
implementation that produces these.