(csskit_derives): clean up derive code (#170)
Some of the csskit_derives code was very hastily/sloppily written. I
took a bit of time to clean it up, tighten some of the code, make it
less brittle and a bit more generic. Some nice benefits of this is that
now it can apply to more structs, so I went through and added
`derive(Peek)` in a few more places that it could be added.