css_ast: derive IntoSpan for all AST Nodes (#174)
It's important that AST nodes have an easy way to get their span. This
is useful in, for example, `csskit_highlight` which takes those spans to
highlight code.

Traditionally we hand wrote `Into<Span>` implementations. However this
is laborious and so they were done sparingly. While cleaning up derive
code in #170 I realised we could probably make a derive macro for this,
and so I did.

So this change adds the new `IntoSpan` derive macro, which will figure
out the first and last members of an AST node and get their spans,
adding them together. The result is a generated implementation that can
reliably output Span information for all AST nodes.

This change also applies `IntoSpan` to all viable AST nodes, by first
applying it to `Stylesheet` and working our way down.