(css_ast): enables `<content-list>` (#175)
To enable `<content-list>` we needed to enable a bunch of other types,
like `<leader-type>`, `<leader()>`, `<content()>`, `<counter>` and more.

I also added blanket implementation for `Vec<\'_, T>` that implements
`Parse` too, which unlocks being able to derive Parse for things like
`struct(Vec<'a, Foobar>)`.

I can see a second pass through that could enable deriving Parse for an
enum of functions, and for something like `Option<T>`. As well as a bit
of a tidy up to make use of `Vec::parse`.

closes: https://github.com/csskit/csskit/issues/154