Implement -moz- pseudo elements, and some blanket impls for Peek/Parse/ToCursors (#176)
I wanted to drop `impl ToCursors for () {` because I'm sure this will
cause us bugs in the future, but to do so we needed to implement
`MozFunctionalPseudoElement`, which relied on this blanket impl. I
thought I'd use it as an exercise to impl some more blanket impls for
the Park/Peek/ToCursors traits, which got us some nice succinct code for
this Parse impl. We're very close to being able to just `derive(Parse)`
on this one, but still need a way to discriminate on `T![Function]` &
`!T[Ident]` & co, so not just yet.