Modify all Into<Span>/From<T> for Span to `ToSpan` trait (#182)
After running into some tricky recursion issues due to implementing
From/Into Span, I realised this isn't the best way to create a trait an
unilaterally apply it to a bunch of generics, due to - I think - From<T>
for Span getting the blanker Into<Span> which can create a recursive
relationship for things like `From<&Vec<T>> for Span where &T: Span`
because `&T` could also be a `&Vec<T>`. After much trial and error it
seems the way around this is just a custom trait, which has less
automatic implementations and therefore subtly avoids these recursion
issues.

Conveniently we already had a macro for this - Spanned - , because
apparently we'd encountered issues before for `From<T> for Span` (in
that it's sized so not dyn compatible). It didn't dawn upon me when
implementing a bunch of `From<T> for Span` impls to use the thing right
in front of my eyes, so I pressed on only to get bit by another issue
with From.

This also changes the name of `Spanned` which could have a clearer name,
so it's now `ToSpan` (akin to ToCursors). This also means a name change
for the derive trait, hence the massive diff of renames. Other than the
renaming, not too much has changed. One additional small refactoring
here though, is using a macro for `ToSpan for (T, U...)` style tuples,
so we can easily add more should we need them.

All this to say hopefully this will avoid recursion issues as we
implement generic structs.