add comma separated struct (#185)
`Vec<'a, (..., T![,])>` is easy enough to write, but `CommaSeparated<'a,
...>` is easier and more readable.

This replaces all _simple_ instances of `Vec<'a, T, Option<T![,]>>` with
`CommaSeparated<'a, T>`. Some of these were using traits just to aid in
writing the parse steps for these, so they've been dropped. Those are:

- CommaSeparatedPreludeList
- SelectorList

The list of types which have been changed to use the new
`CommaSeparated<T>`:

- `SelectorList`
- `DocmentMatcherList`
- `KeyframeSelectors`
- `LayerNameList`