csskit_proc_macro: tidy up def.rs & break out into generate.rs (#197)
def.rs was getting a little unweildy, as it was over 1kLOC. This aims to
break it up a bit by separating the two responsibilities:

- def.rs is now responsible for defining the types, and parsing the CSS
type definition grammaer into Defs.
- generate.rs is now responsible for taking a Def and generating Rust
data structures, such as the struct/enums, as well as the necessary
variants & types.

In moving them over, I've also spent a bit of time cleaning the code.
Some various wins:

- The ToType trait now generates a list of types that each Def
corresponds to. The old code for this used to couple variant names and
the inner types, but that's been massively simplified so that ToType can
be relied on to get a list of types and nothing more.
- ToType also has a `to_singular_type()` which will return either the
single type or the tuple of the types, which makes it easier to use this
type both inside of struct/enums but also as peek/parse steps.
- The ToField trait now encapsulates the variant/member name logic.
- We're using `CommaSeparated<T>` now for `OneOrMoreCommaSeparated`
which is a tiny bit cleaner.

Some things still todo:

- I'm sure more parse/peek steps can be simplified. 
- I haven't yet got codegen working for `Optionals!` as it's tricky due
to the keyword stuff. We should look at this more.