Create Visitable derive macro, derive it everywhere. (#199)
Visitors have have a fun history here. They started out in 907011a, but
got put to bed for a while, revived in 4465ac7, but with lots of todos
and in a way which didn't gel very well with the rest of the codebase.
They hobbled along until PR #64 (b6083c4) where there was an attempt to
revive them so we could get semantic highlighting working in the LSP.
The problem is they were working just well enough for demonstration
purposes but we needed something a little more comprehensive.

This commit iterates again, but this time hopefully more earnestly. This
tweaks the Visitable/VisitableMut traits to be a little more refined by
dropping useless lifetimes, and applies them to some generics like Vec,
Option, CommaSeparated etc. This also introduces the derive(Visitable)
proc macro which will automatically implement both the
Visitable/VisitableMut traits on the given struct.

The `derive(Visitable)` trait steals the `#[visit]` attribute (though
I've left in csskit_proc_macro::visit for now - it's needed for types
that don't derive). `#[visit]` on its own is an alias for
`#[visit(all)]` which will iterate over all members and `accept[_mut]()`
them. Calling `#[visit(skip)]` on any member will cause it to be missed
- which is useful for tokens or other non-visitables. There's also
`#[visit(self)]` for leaf-nodes that want to be visited but don't have
any meaningful children to visit, and `#[visit(children)]` for delegate
nodes that aren't meaningful in of themselves, but can dispatch out to
their children. All in all this makes for a fairly robust way of getting
Visitable derived in almost all AST nodes.

And so this commit _also_ applies Visitable to just about everything. We
should probably put some effort into testing these, but I have some
other ideas around how to make use of these in a unilateral way, so
until then this will have to be considered good enough!