css_ast: tidy up build script (#203)
The build script was getting a little unwieldy because:

- It was all in one file with not much separation going on.
- It has a big old chunk of grep code which was intermingled among debug
lines and cargo build lines.
- It did a bunch of string manipulation to get the matches from grep
into something resembling idents.
- It threw those strings into a bunch of other strings, to write out as
Rust, rather than using something a little more robust like Syn.

This is an attempt to start to fix some of those issues and clean it up
a bit:

- I've split a major chunk into a separate library, which means it can
be reused, as well as better tested/modified/separated further. But
also...
- The grep code is now separated (in the library) from the cargo build
code, and so cargo logs for example aren't quite so tangled within it
all.
- The new library outputs `syn::Type`s rather than strings, so there is
a much better sense that the matches actually parse to valid Rust.
- The build.rs now uses syn/quote to generate Rust code, which gives me
a lot more confidence that it will generate syntactically valid rust.
Also it's much easier to read to boot.

Overall perhaps not _less_ code, but definitely cleaner code.