css_ast: avoid Vec in Nth type (#201)
This change switches the internal representation of Nth to drop the Vec,
instead using a slice of 4 cursors. While this increases the size of Nth
by 12b, it does so avoiding an allocation (because now Nth is a fixed
size) and importantly it doesn't increase the overall size of `Selector`
as a consequence (which is still 128b). This should have always been a
slice rather than a Vec, because we have a known maximum size of cursors
that can possibly be expressed in an Nth.

This change requires some kind of sentinel "Empty" Cursor, which is a
little trickier. Originally I thought I'd use an Option<Cursor> but that
inflated the size by another 20b because Cursor & Token are already
tightly packed. Luckily it's not _that_ tightly packed, that we couldn't
carve out a little space: a Token can be Whitespace, with the Whitespace
bitflags all set to 0, this can effectively represent a zero width
token.

A slight downside to all this is that the serde serialisation gets more
junky, but it is already bad enough so this doesn't make it
significantly worse.