css_ast: support url()/src() functions in images
`<image>` has a syntax of `<url>`. `<url>` is a type, distinct to `<url-token>` (the token). The `<url>` type is one of
`url()` or `src()`. This type also allows for either the `<url-token>` production or cases which do not satisfy that lex
token, and instead lex as a Function.

This change introduces the new `<url>` type, and refactors `<image>` to use this type.