csskit_derives/css_ast: allow state/stop to be set in derive(parse) (#211)
In the interests of optimising properties/mod.rs, I spotted there was
some repetition in the Custom/Computed/Unknown parse steps, this is
mostly because each one needs to set state and stop tokens. If only
derive(Parse) could somehow do this then we could derive(Parse) on all 3
of these.

So this change does exactly that! This introduces the
#[parse(state=...,stop=...)] syntax, which allows derive(Parse) to issue
some before/after steps to set the state, and tear it down when done.
This is not only reduces code in css_ast but also should make for a more
consistent & bugfree code for any other property that wants to use
state/stop because it can be applied correctly in the derive.