csskit_proc_macro: Generate empty CommaSeparated with new_in, not Default (#252)
In https://github.com/csskit/csskit/pull/248 I mistakenly added steps
for generating <type>#? values by using
`Default::default()` for CommaSeparated, which is not possible because
it relies on the Bump heap allocation.

This change adds `new_in` method for CommaSeparated, and uses that in
the generate code so that this branch actually
works and isn't a compile failure.