optimise Alternatives of auto|<length(-percentage)> to <length(-percentage)-or-auto>
Syntax such as `auto | <length>` appears in quite a few values, enough for us to have already created the `LengthOrAuto`
and `LengthPercentageOrAuto` types which simplify this. However def.rs/generate.rs have not supported these types
traditionally.

This change alters def.rs to support these types, by checking for these specific cases - a combinator of alternatives
with a length of 2 where one alternative is the `auto` ident and the other is either `<length>` or `<length-percentage>`. This is a very specific optimisation but one that can be somewhat broadly applied, so is worth implementing.