csskit_proc_macro: Remove DefGroupStyle::Range/OneMustOccur, use Def::Multiplier instead (#256)
The implementation of Def originally stuck quite close to css-values'
[component value modifiers][1], but as we get
close to covering the whole grammar, things become more readily
apparent, such as the work to optimise away groups done
in https://github.com/csskit/csskit/pull/243 - which tried to make it so
that DefGroupStyle::None was never a case in
the generated code. Another such win is also possible - the
DefGroupStyle::OneOrMore and DefGroupStyle::Range are
(effectively) semantically identical to Multiplier of
OneOrMoreCommaSeparated and Multiplier with a Range.

So this change removes the DefGroupStyle::OneOrMore and
DefGroupStyle::Range variants, and instead during definition
parse time those groups will be converted into the respective Multiplier
values. The result of this should be that
generator code no longer needs to worry about Groups (apart from the
OneMustOccur group) and instead the Multiplier
branches can cover these cases.

[1]: https://drafts.csswg.org/css-values-4/#component-combinators