csskit_proc_macro: split out optimization pass of Def parsing, optimize more
Following up from https://github.com/csskit/csskit/pull/256, a syntax like `[ auto | length ]{1,2}` could be optimised
as a Multiplier of LengthOrAutos, but we can go a step further and optimize it to a Combinator of [LengthOrAuto,
Optional<LengthOrAuto>]. We already do this throughout the parsing phase when other Multipliers are generated, but
https://github.com/csskit/csskit/pull/256 introduced a new callsite where a Multiplier could be generated.

This change optimizes that new case, but rather than repeating code it takes the opportunity to pull out the
optimization steps into a separate function which can be repeatedly called to help minimise the depth of a parse tree.
Practically speaking this really only benefits this case for now, but this gives us an easy place to further add
optimisations without having to entangle them further in the parse steps.