Implement TimeOrAuto, optimize for it in Defs, and uncomment AnimationDurationStyleValue.
In https://github.com/csskit/csskit/pull/257 we optimized for values like `auto | <length>` to use the LengthOrAuto
struct, but we can also do the same for syntaxes that use `auto | <time>`.

This change introduces a TimeOrAuto enum which is similar to LengthOrAuto. This change also implements the requisite
code paths in def.rs and generate.rs to leverage this new type, which lets us unlock animation-duration!