csskit_proc_macro: Generate structs for multipliers of keywords
Some style values are just a multiplier over a keyword set. These can be somewhat easily generated by using the
`FooKeywords` type which we already generate for when there are many keywords.

So this change does that, we can do a little trick by matching for these cases (a multiplier of a combinator of
alternatives where all the alternatives are Idents) and generating a _new_ Def which is a multiplier of the type
`FooKeywords` (where `Foo` is the struct name). This then generates the appropriate parse/peek code.

This should unlock extra values.