generate-values: refine heuristics for enum/struct determination (#272)
The enumOverrides/structOverrides maps were getting a little large, and
so I decided to take a look at the way we
determine if a type should be an enum or struct, and work on reducing
the overrides map. The result is that the maps are
now empty, because we can reliably determine the type using heuristics
alone. The maps will remain because they may be
useful for future properties but I think this puts us in a better place
for value generation.