csskit_proc_macro: Generate sub-types where applicable (#273)
Many style values are a multiple of some combinator, where the easiest
way to generate the value would be to use a
subtype. Some grammars explicitly do this, for example
`<single-animation>#`, while others are more implicit, e.g.
`[ none | <keyframes-name> ]#`. For these multipliers, we can generate
the additional sub-types during the proc macro
phase, much like we do with the `*Keywords` structs.

This change introduces a `generate_additional_types` method to `Def`,
which takes the existing `generate_keyword_set`
function and enhances it, adding sub-types where needed. Also
complemented with a snapshot test which looks like it's
generating the right code!

With this change a grammar like `[ none | <keyframes-name> ]#` on a
struct like `AnimationName` will generate a new
type of `SingleAnimationName` with a grammar of `none |
<keyfrmas-name>`. The generation for `AnimationName` is then
trivial as it becomes a `CommaSeparated<SingleAnimationName>`.