Implement combined types NumberPercentage, NumberLength, NumberLengthOrAuto.
Many types are based on a combination of Length/Percentage/Number, such as:

- `<length> | auto`
- `<length> | <percentage>`
- `<number> | <percentage>`
- `<number> | <percentage> | auto`

For many of these types we - and CSS - flattens them into more convenient types such as `<length-percentage>`. Some,
however, are not.

This change finds more of these types and collapses them into such combined-types, namely:

- `<number> | <percentage>`
- `<number> | <length>`
- `<number> | <length> | auto`

This should unlock more properties for us to enable.