css_parse: impl CursorSink for Vec
For a while we've had a CursorVecSink which acts as a simple way to collect a set of cursors into a Vec. This, however,
was a little naive, as the trait can simply be implemented _for_ Vec.

This change does just that - impls CursorSink for Vec, thereby making CursorVecSink redundant.