css_parse: implement CursorOverlaySink
We need some way to be able to overlay ASTs on top of one another, so we can start applying transforms again. One way to
do this would be to introduce a mechanism for overlaying two CursorSinks on top of one another, or to have an AST be able to be replayed on top of a delegate CursorSink.

This changes introduces CursorOverlaySink: a new CursorSink which takes an underyling SourceCursorSink and feeds it
Cursors that get appended, unless there's an "overlay" - a set of cursors ready to replace the cursors within a given
span. This allows us to take N+1 ASTs and smoosh them together, such that the base AST gets new ASTs pasted over the top
in a way that produces a combined AST.