csskit_proc_macro: Visit sub-types of StyleValues
When generating StyleValues in https://github.com/csskit/csskit/pull/318, the easy choice was to make them
`#[visit(self)]`, but of course that's a stop gap, as we want to be able to visit individual types, such as functions or
units. https://github.com/csskit/csskit/pull/321 added sufficient Visitable capability to all the necessary types, such
that now we can drop `#[visit(self)]` on StyleValues and instead use `#[visit]`, with an occasional smattering of
`#[visit(skip)]` on the sub-types.