css_lexer: Allow String tokens to change quote style.
Formatters/serializers might want to change the quote style of a given String. We should allow this at the lexer/cursor level.

This change adds a new `with_quotes()` method on Token, which returns a new token with the given QuoteStyle. This also
makes changes to Cursor::write_str to ensure that it looks at the QuoteStyle and writes the correct char, ignoring the
underlying backing string's quote style.

This also has the side effect that strings without a closing quote will have their closing quote written out. This is
actually worthwhile doing because the only valid String tokens where a closing quote isn't present are those that close
out with an EOF. They're not really encountered in the wild as CSS that ends the file with a string invariably contain
junk.