css_ast: Introduce AutoOr<T>, NoneOr<T>, AutoNoneOr<T> types.
CSS has a whole boat load of style values or sub-types which boil down to `auto | <x>` or `none | <x>`, and even
occasionally `auto | none | <x>`.

This change introduces this concept as a generic type. This is quite a big change because of the following incidentals:

- Changing a type from `auto|<x>` to `AutoOr<X>` changes it from an Enum to a Struct, so some modifications to
	generate-values needed to happen.
- The generator required some substantial changes to support these types
- We still want checks to be performed on a value like `auto|<percent [0,100]>`, which furthered the work in the
	generator.

During the work on checks I took the time to look at all check code in the codebase and noticed some other places that
needed fixups. I should split those out but it's late...