css_ast: Use #[syntax] for AnchorName.
Following https://github.com/csskit/csskit/pull/349 & https://github.com/csskit/csskit/pull/350 we can now start to use #[syntax] for types as well as values.

This change dips our toes into this by implementing this for AnchorName - which is a simple type which wraps a
DashedIdent. This change also adds relevant tests for this type to ensure it works correctly.