csskit_proc_macro/csskit_derives: Implement AllMustOccur simple cases
The AllMustOccur combinator style wasn't parseable in #[syntax], erroring out. This change introduces some very
basic tests and the basic parsing code in order to generate an AllMustOccur style syntax implementation. In addition to
that, in order to try to keep parity with derive(Parse) this introduces the #[parse(all_must_occur)] attribute in
struct/enum variants for derive(Parse) to achieve the same functionality.