css_feature_data: Add new css feature data!
In https://github.com/csskit/csskit/issues/338 @foolip introduced us to the [baseline & web
features](https://web-platform-dx.github.io/web-features/) repo which houses most of the data we need for web features,
and does so in a single JSON file which is arguably more convenient and more robust than pulling metadata from the website for each
feature. @foolip also suggests to build a BCD key for each feature and use that to get the relevant metadata such as
platform support & so on.

This (large) change introduces a new package: css_feature_data, a generated data package which takes the CSS Feature
data from Web-Features and puts it in a rust-friendly structure, optionally including `browserslist-rs` support. This
package has its data generated using `generate-all-values` meaning it'll be updated weekly along with the various
values.

This change also incorporates css_feature_data into css_ast by introducing a new csskit_derive::ToCssFeature which takes
the key name as an annotation and produces a ToCssFeature trait impl for the type, using the key to fetch the relevant
feature. Lastly, to test this I've introduced an assert_feature_id! test helper to allow us to validate that the various
features are being used.

This should cover a lot of the features in the package automatically, due to style value generation. The derive also
attempts to pull out the right BCD key for enum variants which seems to work reasonably well. We'll see how well this
continues to work.

I'd say this gives us a more solid foundation for using BCD and baseline, so I'm going to consider this change as one that closes https://github.com/csskit/csskit/issues/338.