css_parse: Allow CursorPrettyWriteSink/CursorCompactWriteSink to push to other sinks.
Rather than these sinks pushing to a fmt::Write, they can be more flexible by pushing their tokens to another
SourceCursorSink. We already have an implementation of SourceCursorSink for String, so most of the use of these is
already handled for free, but it means we get to wrap them in more interesting ways.