css_lexer: Fix hex_value parsing
https://github.com/csskit/csskit/pull/399 tried to eagerly parse hex codes and stuff it in token data, but the
implementation was naive and quite broken. It collected the hex characters but did not "pad" for shorter values such as
when the alpha channel was missing (in 3 and 6 character hex codes) or when the letters were "short form" (in 3 and
4 character hex codes).

This changes fixes the parsing to be more accurate, so that the resulting hex values can actually reproduce the
underlying hex code.