chromashift: Introduce chromeashift
chromashift is a package which can represent many (soon to be all) of the CSS colour formats, as well as the mechanisms
and algorithms defined by css-color. It includes the ability to convert between all colour types using From/Into, and
has the ability to mix colours using CSS mixing algorithms & ergonomics.

In addition to trying to hit all of css-color-5, it also includes WCAG contrast to determine WCAG contrast ratios and
find colours that meet those requirements, and utilities to see how close one colour is to another using CIE Delta.

This should provide a good foundation for all the required colour math needed for things like minification of colours.