drop Build trait
This trait was supposed to provide a signal for when a node could be infallibly parsed from a single cursor, but in
reality it was tied to it's peek implementation (and in fact many times we ended up adding debug_asserts to try to
ensure that peek was properly called before a build). All-in-all the trait wasn't that helpful and it wasn't that much
easier to implement Build than simply implementing Parse.

This change drops the Build trait, and the work cascades from there. All types that used Build now use Parse instead,
which highlighted a few bugs here and there - so they've been fixed. Some structs were simple enough to drop Build and
derive(Parse), but sometimes that highlighted subtle issues to fix. This work also lead to dropping the
`T![Dimension::*]` types as they mostly provided convenience for building but got in the way, so using `T![Dimension]`
made life simpler in a Parse-trait only world.