Replace miette diagnostics with custom Sized struct.
One issue we've had with the Parser is that it performs allocations on the heap to build errors & trivia. This is
effectively unavoidable due to the variable amount of these, but we use the bumpalo memory arena to try and amortize
allocation costs. The only issue is that historically the Trivia & Error Vecs _haven't_ used the arena. For Trivia, this
is mostly incidental, but for Errors this is due to the fact that Miette Boxes its Error type so it can be dyn (I
think?). The result is that even if we used Bumpalo's Vec, we'd still be heap allocating.

This is made worse because some parser steps will iterate through a Parse step only to produce an Error, (therefore
allocate) only to rewind, discard the newly allocated error, and branch down a different path to not error, meaning a
complete AST can be built whilst allocating data that didn't need to be allocated!

This rather unfortunately large change alters our Diagnostics such that they're defined in a compact struct which is
easy to build (it comprises of a few cursors and a function pointer) and it can be Sized, meaning no heap allocation to
_build_ an error, which means Parse results can be allocation free! This and this allows us to very easily move the
allocation into the errors Vec to use a Bumpalo Vec instead, relying on the pre-allocated Arena instead of potential new
allocations on each push.

This change sees a modest improvement in some parse times, likely because the files being parsed are hitting errors.
This change wasn't done so much to see parse time improvements on its own, but is part of an (even larger) set of
changes around the parser mechanics...