css_ast: tidy up mod/impl generation
This changes `values/mod.rs` to include a `prelude` mod which all values/**/mod.rs files now use, instead of relying on
copy-pasting a bunch of stuff in each `impl.rs` file. The `impl.rs` files are now largely for tests, and the occasional
extra required implementations for things.

This also updated to `web-features@next` and tweaks the generator to accomodate the new schema.