csskit_derives: Add support for deriving on generics
The AutoOr/AutoNoneOr/NoneOr types had quite a bit of boilerplate purely because they needed to impl
Parse/Peek/ToSpan/ToCursors/Visitable/VisitableMut for T where T also impld that trait. The core missing puzzle piece
being that the derives of those traits did not support deriving where the generic needed to also impl the trait - you'd
just get a compile error saying the root type needed to enforce that trait bound, which was not desirable as we only
want to, for e.g., derive Visitable for T when T is _actually_ visitable.

This change incorporates this capability into csskit_derives such that if a type with a generic is derived, then the
automatically derived implementation will add a where clause saying that each of the generics needs to implement the
trait, but only when the generated actually calls for it.

This change then also "self-hosts" the change by making the AutoOr/AutoNoneOr/NoneOr types all derive these traits,
removing quite a bit of said boilerplate.