css_parse: implement ordered cursor sink for trivia output
In https://github.com/csskit/csskit/pull/499 we detailed an issue with current ToCursors implementation which is that
cursors may arrive out-of-order. This is largely to do with CSS' native grammar ordering, which is not always the
authored order. This is important to retain, but for IDEs this is undesirable; code actions should not re-order authored
text!

This change adds CursorOrderedSink, a sink that buffers and emits cursors in source order to preserve original
formatting. This tries to efficiently buffer cursors until a contiguous region of buffered cursor data is ready, which
it will then emit in source order.

This change also incorporates CursorOrderedSink into `assert_parse!`, negating the need for a second "expected" argument
for the macro, which was mostly used to re-interpret inputs which are not in grammar order. Now `assert_parse!` always
asserts in source order, instead.

Some tests are failing because they use trivia (comments). Those (small number of) tests have been commented out
temporarily, we'll add trivia output to `assert_parse` shortly.