css_lexer: Allow custom allocators, drop bumpalo as a core dependency.
While we're unlikely to move away from Bumpalo any time soon, it would be nice to improve the code to use the more
generic/"idiomatic" nightly allocator APIs, and we can do so by using allocator-api2.

So this change alters css_lexer to move away from requiring Bumpalo to parse string escape data, and instead rely on the generic Allocator trait. This should allow us to shift to different allocators if the need ever arises, without a huge migration later.

This also simplifies tests a tiny bit, as we don't care so much about allocations during (most) tests and can simply
allocate using the global allocator for these.

This is the simplest package to do this in, as these are the only two places that allocate, but it does mean moving
these methods to no longer return Box str shenanigans and moving to using a proper CowStr implementation, which
unfortunately we had to hand roll due to std Cow not implementation custom allocators, and allocator-api2 not
implementing a Cow.