CI: ensure website builds on main, move benchmark-compare to nightly.
Having benchmarks run on every commit to main is a little silly considering we commit a lot per day. But also the
current build means the website is a manual deploy job (because actions commits don't kick off other actions).

Instead this change moves the website back to building on main, and moves the benchmark job to a nightly one. The
website also builds nightly, just in case we haven't made any commits to main but we want fresh benchmark data.