CI: try to use benchmark-tracking for PR benchmarks
Our PR benchmarking scripts use a criterion-compare action which is useful, but slow because it runs the whole suite
twice - once for main, then again for the PR. Considering we benchmark nightly, we could use that data to build a table.
In addition the nightly job runs more benchmarks, e.g. hyperfine. We can benefit from this in PR stats.

This change downloads the benchmark-history artifact and runs the full bench suite against it, then compares the latest
result (the PR) with the 2nd entry (last nightly run) to determine the delta.