css_ast: Gate Visit behind a feature flag
The Visitable feature generates quite a lot of code which contributes to:

 - Slower build times
 - Difficulty debugging sometimes

In order to improve this, I set about making Visit gated behind a `visitable` feature flag. It was... a lot. I kind of
regret started it given how many files it touched, but it's done now and I'm not undoing it.