css_parse: Ensure trailing/leading descendant combinators aren't parsed
Currently in certain cases the descendant (whitespace) combinator will be parsed if it's at the start of end of a
selector, which is incorrect. This is because the parser checks based on the current skip state which could include
whitespace.

This change specifically sets the state while peeking at the next token, to ensure that if it peeked a stop token it can
break rather than consume & parse the whitespace.