css_ast: Re-work CssAtomSet to allow for vendor prefix bitflags
Prior to this change CssAtomSet did nothing special with vendor prefixed atoms,
which was fine, but they have intrinsic information about them which we can use.

This change re-orders and re-values the vendor prefixed atoms so that they're at
the bottom of the CssAtomSet enum, and they set the highest bit (23) to signify
they're a vendor prefixed atom. In addition each vendor gets its own bit among
the 21st & 22nd bits, which allows us to trivially convert these two bits into
an index to lookup the vendor prefix; making for a trivial TryFrom for
VendorPrefixes.

This then ensures all properties set the vendor_prefixes metadata.