css_parse/css_ast: Improve NodeWithMetadata to allow extracting self meta
Prior to this change NodeWithMetadata had two methods: self_metadata and
metadata, without a particularly clear distinction on when to use either.

This change alters NodeWithMetadata so that self_metadata is more clearly just
the metadata that the Node itself would alter, while metadata is the Node's
metadata plus any child metadata, merged.

This change also goes and updates all implementations to follow this pattern -
so self_metadata generates the nodes own metadata.