csskit_ast: Improve :not & sibling selector matching
Prior to this change deferred paths would be silently ignored due to left-hand
siblings or metadata dependent pseudo selector matching being skipped.
Additionally :not() only checked the inner right-hand type, meaning it almost
always failed.

This change adds a bunch of tests for various combinations of sibling & not
selectors, and fixes the issues by passing around MatchContext in SiblingInfo,
and evaluating the whole selector inside of `:not()`.