css_lexer: Add DynAtomRegistry
While CssAtomSet is able to intern all pre-defined keywords, css sheets often
have a lot of custom keywords, for example `.class` selectors or
`--dashed-idents`. To make parsing and comparison of these more efficient, it
would be nice if we could intern them somehow. We could add a whole bunch of the
most popular to making parsing and comparison more efficient for popular
libraries, but this is very brute force, very manual, and doesn't solve for all
use cases. (We did this a while ago naively trying to speed up parsing).

Looking at the string-cache crate (https://github.com/servo/string-cache), which
this project originally used, it retained a static atom set, and also a dynamic
atom set. The dynamic atoms would effectively "leak" strings into a dynamic
registry that would build up during runtime. This is generally okay to do
because the strings we intern are keywords (as opposed to string data) and so
they're mostly very small and repeat a lot, but also worth retaining for the
lifetime of the program.

So this tries to re-implement the same principle, albeit in a different way.
DynAtomRegistry is a new wrapper around an AtomSet which implements DynAtomSet.
As atoms are ingested if they're not in the given AtomSet, then they're leaked
(Box.leak) and interned into a u32. This means DynAtomRegistry never returns
`_None`, and instead returns a new u32 for each new string. Static Atoms have
bit 23 (the MSB in token packed data) set to 0 while dynamic atoms set it to 1,
ensuring that static vs dynamic atoms are still distinguishable from each other.
This reduces the effective amount of static tokens from 16MM to 8MM (so still
essentially unlimited) and allows for 8MM dynamic atoms.

Each AtomSet (CssAtomSet, CsskitAtomSet) gets a singular registry which is
statically allocated (via LazyLock). The dynamic registry uses fnv to hash the
value. If the hash misses then a Vec is pushed to and the new `len()` becomes
that atoms u32 (so first dynamic atom is u32 `1`, second `2` and so on). Hash
hits will return the prior u32. This method of using the Vec makes it easy to
get the `&static str` back from an atomized string.

The nice thing about wrapping DynAtomRegistry is it makes it entirely _opt in_,
so we only need to use it in the select cases where it is actually beneficial to
leak/register these.